requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# AWS Services
boto3==1.34.34
botocore==1.34.34

# Data & Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Utilities
tenacity==8.2.3  # Retry logic
httpx==0.26.0  # HTTP client for testing

# Monitoring & Logging
watchtower==3.0.1  # CloudWatch logging
sentry-sdk[fastapi]==1.40.0  # Error tracking

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1  # Code formatting
flake8==7.0.0  # Linting
mypy==1.8.0  # Type checking

# Optional: Frontend (if using)
# nextjs dependencies would go in frontend/package.json
